Matrix Calculator, Version 1
First version of a matrix calculator which uses Gaussian reduction to solve a matrix. 
As this is a first version, it works properly only in the case of square matrices with a unique solution. (It works for those with no solution too, but is up to the user to notice the inconsistency.)
More efficient methods and a more generally working program will be implemented in time.